strata test env e2e test refactor
