+++ strr-strata-web/tests/e2e/enums/login-source.ts
@@ -1,0 +1,4 @@
+export enum LoginSource {
+  BCSC = 'BCSC',
+  BCEID = 'BCEID'
+}
