+++ strr-strata-web/tests/e2e/test-utils/generate-otp.ts
@@ -1,0 +1,11 @@
+import * as OTPAuth from 'otpauth'
+export function generateOTP (secret: string) {
+  const totp = new OTPAuth.TOTP({
+    secret,
+    digits: 6,
+    algorithm: 'sha1',
+    period: 30
+  })
+
+  return totp.generate()
+}
